This is a big one. I am porting HTTP server to use CAIRO storage system. This turned out to be a complete rewrite. I just started testing ported code to realise the number of gaps, bugs and inconsistent behavior in code that imports text files into database. I am aiming to get test coverage to 100% and test all use cases I can think of right now. I am not far off the target but there are few more things still to be completed. Here is summary of the change:
- ported text file parser and import mechanism to CAIRO
- added a tonne of tests to get code coverage to 87% for now
- fixed a number of bugs with encoding two 32-bit integers as single 64-bit one
- consolidated table creation code (there were 4 copies, now there is one working consistently)
- copied HTTP server into "tuck" package, which is where new server will be evolving before it moved elsewhere

+++ core/src/main/java/com/questdb/std/FilesFacade.java
@@ -77,7 +77,7 @@ public interface FilesFacade {
 
     long openRW(LPSZ name);
 
-    long read(long fd, long buf, int size, long offset);
+    long read(long fd, long buf, long size, long offset);
 
     boolean remove(LPSZ name);
